feat: DMATH-52 убираем лишние описание
