+++ src/app/algorithms/hungarian/page.tsx
@@ -187,33 +187,6 @@ function HungarianContent() {
         </Box>
       </Paper>
 
-      <Paper
-        sx={{
-          p: { xs: 2, sm: 3 },
-          background:
-            'linear-gradient(to bottom right, rgba(59, 130, 246, 0.1), rgba(147, 51, 234, 0.1))',
-          border: '1px solid rgba(59, 130, 246, 0.3)',
-        }}
-      >
-        <Typography variant="h6" sx={{ mb: 2, fontWeight: 600, color: 'info.light' }}>
-          Как использовать
-        </Typography>
-        <Box component="ol" sx={{ pl: 2.5, display: 'flex', flexDirection: 'column', gap: 1 }}>
-          <Typography component="li" variant="body2">
-            Введите квадратную матрицу стоимостей (числа — стоимости назначений)
-          </Typography>
-          <Typography component="li" variant="body2">
-            Нажмите "Отправить" чтобы построить двудольный граф
-          </Typography>
-          <Typography component="li" variant="body2">
-            Нажмите "Запустить" чтобы выполнить алгоритм
-          </Typography>
-          <Typography component="li" variant="body2">
-            Просмотрите результаты: оптимальное назначение и общая стоимость
-          </Typography>
-        </Box>
-      </Paper>
-
       <Alert
         open={alertState.open}
         onClose={closeAlert}
